refactor!: remove unstable/low_level, harden pre-1.0 surface (M6)
Design-hardening pass before the 1.0 surface becomes a semver-locked
contract. Pure visibility/organization refactors — detection results are
bit-stable; all gates green, WASM builds, Python 86/86 (incl. .pyi parity).

BREAKING CHANGE: removed `chess_corners_core::unstable` and
`chess_corners::low_level`.

- DEBT-01: delete `chess_corners_core::unstable`. The root-public
  chess_response_u8 / find_corners_u8 / detect_corners_from_response
  required `ChessParams`, which lived only in `unstable` under a "may
  disappear in any patch" disclaimer — the stable contract was uncallable
  without a no-semver type. Promote ChessParams, RefinerKind,
  chess_response_u8_patch, detect_peaks_from_response_with_refine_radius,
  refine_corners_on_image to the documented crate root; demote the ring
  tables, ChessParams::ring(), MAX_IMAGE_UPSAMPLE, and
  find_corners_u8_with_refiner to pub(crate); make chess_response_u8_scalar
  test-only golden infra; delete the dead detect_peaks_from_response.
  White-box tests relocated from tests/ into in-crate unit modules.

- DEBT-02: delete the facade `chess_corners::low_level` escape hatch (no
  real hand-composer consumer). Expose lowering as public
  DetectorConfig::{chess_params, radon_detector_params, coarse_to_fine_params};
  hand-composers depend on chess-corners-core directly.

- DEBT-03: retire the too_many_arguments #[allow] cluster on the
  orientation/detector hot paths — 4 fixed by collapsing (img,w,h) into the
  existing Copy ImageView, 3 were stale and removed. No unjustified #[allow]
  remain in scope; stack-only, no per-corner alloc, bit-exact.

- DEBT-04: replace the disk-sector argmax -1.0f32 sentinel with Option.

- DEBT-05: split the 1008-line facade config.rs into a cohesive config/
  module (public paths byte-identical; lib.rs pub use unchanged).

- docs: record M6 + DEBT-01..08 in ROADMAP/BACKLOG, annotate the superseded
  API-03 RFC, add a release-readiness banner (no P0/P1 blocks the freeze).

Migration: see CHANGELOG.md [Unreleased]. Lower a DetectorConfig with
DetectorConfig::chess_params() etc., or depend on chess-corners-core for the
stage functions.
