+++ crates/chess-corners/src/config/mod.rs
@@ -1,0 +1,12 @@
+mod chess;
+mod detection;
+mod multiscale;
+mod radon;
+
+#[cfg(test)]
+mod tests;
+
+pub use chess::{ChessConfig, ChessRefiner, ChessRing};
+pub use detection::{DetectionParams, DetectionStrategy, DetectorConfig};
+pub use multiscale::MultiscaleConfig;
+pub use radon::RadonConfig;
