+++ crates/chess-corners/src/diagnostics.rs
@@ -76,7 +76,7 @@ impl<'a> DetectorDiagnostics<'a> {
     /// [`ResponseMap`] therefore has the same dimensions as the input.
     ///
     /// This mirrors the free function [`chess_response_u8`] but sources
-    /// [`ChessParams`](chess_corners_core::unstable::ChessParams) from the
+    /// [`ChessParams`](chess_corners_core::ChessParams) from the
     /// bound detector.
     ///
     /// # Errors
