+++ crates/chess-corners/tests/refiner_benchmark.rs
@@ -15,9 +15,7 @@
 //!
 //! Without `--all-features` the ML column is omitted.
 
-use chess_corners::low_level::ImageView;
-use chess_corners_core::chess_response_u8;
-use chess_corners_core::unstable::ChessParams;
+use chess_corners_core::{chess_response_u8, ChessParams, ImageView};
 use chess_corners_testutil::{aa_chessboard, add_gaussian_noise, gaussian_blur};
 
 #[path = "common/refiner_harness.rs"]
